requirements.txt
requests
alwaysdata_api
dnspython
